+++ lib/ember_cli/sprockets.rb
@@ -1,17 +1,17 @@
+require "ember_cli/errors"
 require "non-stupid-digest-assets"
 require "ember_cli/html_page"
 
 module EmberCli
   class Sprockets
+    class AssetPipelineError < BuildError; end
     def initialize(app)
       @app = app
     end
 
     def register!
-      assets = %r{\A#{app.name}/}
-
-      Rails.configuration.assets.precompile << assets
-      NonStupidDigestAssets.whitelist << assets
+      register_or_raise!(Rails.configuration.assets.precompile)
+      register_or_raise!(NonStupidDigestAssets.whitelist)
     end
 
     def index_html(head:, body:)
@@ -39,6 +39,20 @@ module EmberCli
     def package_json
       @package_json ||=
         JSON.parse(app.paths.package_json_file.read).with_indifferent_access
+    end
+
+    def asset_matcher
+      %r{\A#{app.name}/}
+    end
+
+    def register_or_raise!(precompiled_assets)
+      if precompiled_assets.include?(asset_matcher)
+        raise AssetPipelineError.new <<-MSG.strip_heredoc
+          EmberCLI application #{app.name.inspect} already registered!
+        MSG
+      else
+        precompiled_assets << asset_matcher
+      end
     end
   end
 end